SparkMergeTreeWriteSettings
